--- a/docker-compose.yml
+++ b/docker-compose.yml
@@ -47,7 +47,7 @@ services:
       - REACT_APP_GITEA_URL=http://localhost:3000
       - REACT_APP_API_URL=http://localhost:8000
     ports:
-      - "3001:3000"
+      - "3001:3001"
     depends_on:
       - gitea
       - qubots-api