--- a/web_interface/Dockerfile
+++ b/web_interface/Dockerfile
@@ -1,5 +1,5 @@
-# Multi-stage build for React app
-FROM node:18-alpine as build
+# Multi-stage build for React app with Vite
+FROM node:20-alpine as build
 
 WORKDIR /app
 
@@ -7,7 +7,7 @@ WORKDIR /app
 COPY package*.json ./
 
 # Install dependencies
-RUN npm ci --only=production
+RUN npm ci
 
 # Copy source code
 COPY . .
@@ -19,13 +19,15 @@ RUN npm run build
 FROM nginx:alpine
 
 # Copy built app to nginx
-COPY --from=build /app/build /usr/share/nginx/html
+COPY --from=build /app/dist /usr/share/nginx/html
 
 # Copy nginx configuration
 COPY nginx.conf /etc/nginx/conf.d/default.conf
 
 # Expose port
-EXPOSE 3000
+EXPOSE 3001
+
+CMD ["nginx", "-g", "daemon off;"]
 
 # Start nginx
 CMD ["nginx", "-g", "daemon off;"]