--- a/web_interface/src/App.tsx
+++ b/web_interface/src/App.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { Toaster } from 'react-hot-toast';
 import Layout from './components/Layout';
 import WorkflowDesigner from './components/WorkflowDesigner';
@@ -7,10 +8,21 @@ import ComponentLibrary from './components/ComponentLibrary';
 import Settings from './components/Settings';
 import Dashboard from './components/Dashboard';
 
+// Create a client
+const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      staleTime: 5 * 60 * 1000, // 5 minutes
+      retry: 1,
+    },
+  },
+});
+
 function App() {
   return (
-    <Router>
-      <div className="App">
+    <QueryClientProvider client={queryClient}>
+      <Router>
+        <div className="App">
         <Toaster 
           position="top-right"
           toastOptions={{
@@ -44,8 +56,9 @@ function App() {
             <Route path="/settings" element={<Settings />} />
           </Routes>
         </Layout>
-      </div>
-    </Router>
+        </div>
+      </Router>
+    </QueryClientProvider>
   );
 }