--- a/web_interface/src/components/__tests__/App.test.tsx
+++ b/web_interface/src/components/__tests__/App.test.tsx
@@ -1,0 +1,41 @@
+import { describe, it, expect } from 'vitest'
+import { render, screen } from '@testing-library/react'
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
+import { BrowserRouter } from 'react-router-dom'
+import App from '../../App'
+
+// Create a test query client
+const createTestQueryClient = () => new QueryClient({
+  defaultOptions: {
+    queries: {
+      retry: false,
+    },
+    mutations: {
+      retry: false,
+    },
+  },
+})
+
+const renderWithProviders = (ui: React.ReactElement) => {
+  const testQueryClient = createTestQueryClient()
+  
+  return render(
+    <QueryClientProvider client={testQueryClient}>
+      <BrowserRouter>
+        {ui}
+      </BrowserRouter>
+    </QueryClientProvider>
+  )
+}
+
+describe('App', () => {
+  it('renders without crashing', () => {
+    renderWithProviders(<App />)
+    expect(document.body).toBeInTheDocument()
+  })
+
+  it('has the correct title', () => {
+    renderWithProviders(<App />)
+    expect(document.title).toBe('Qubots - Visual Optimization Workflows')
+  })
+})